Measure more time to let time not be zero
Because originally test function just test one time and if the
computation is simple the measure time will be zero. However, the
program will cut the top slower and faster, due to above, many close to
zero will be zero and finally be dropped, so the result will bigger than
real.
To solve this, we compute 100 times in a time measure. In addition, we
can reduce the test times.